+++ test/destroy.test.js
@@ -20,7 +20,7 @@ test('merge end propagation', () =>
     const res = []
     const s1 = _(h.randomStringGenerator(Infinity))
     const s2 = _(h.randomStringGenerator(Infinity))
-    const s3 = _([s1, s2]).merge(2, false)
+    const s3 = _([s1, s2]).merge({ concurrency: 2, ordered: false })
     const s4 = h.getSlowWritable(res, 1, 10)
     s3.toNodeReadable().pipe(s4)
     s2.once('end', () => {
@@ -39,7 +39,7 @@ test('fork end propagation', () =>
     const s = _(h.randomStringGenerator(Infinity)).makeAsync(10)
     const s1 = s.fork().take(3)
     const s2 = s.fork()
-    const s3 = _([s1, s2]).merge(2, false)
+    const s3 = _([s1, s2]).merge({ concurrency: 2, ordered: false })
     s3[kResume]()
     s1.once('end', () => {
       expect(s.ended).toBe(false)
@@ -90,7 +90,7 @@ test('standard end propagation', async () => {
       s3Ended = true
       expect(s4Ended).toBe(false)
     })
-    .merge(1)
+    .merge({ concurrency: 1 })
     .on('end', () => {
       expect(s1Ended).toBe(true)
       expect(s2Ended).toBe(true)
@@ -132,7 +132,7 @@ test('explicit end', async () =>
         expect(s4Ended).toBe(false)
       })
 
-    const s2 = s.merge(1).on('end', () => {
+    const s2 = s.merge({ concurrency: 1 }).on('end', () => {
       expect(s1Ended).toBe(true)
       expect(s2Ended).toBe(true)
       expect(s3Ended).toBe(true)